requirements.txt
apify-client
chromadb
langchain
openai
python-dotenv
streamlit
tiktoken